+++ src/Controller/Wish/Subscription/WishSubscriptionController.php
@@ -6,45 +6,48 @@ use App\Controller\Controller;
 use App\Entity\Wish\Subscription\WishSubscription;
 use App\Entity\Wish\Wish;
 use App\Entity\Wishlist\Wishlist;
+use App\Security\Voter\Wish\Subscription\WishSubscriptionVoter;
+use App\Security\Voter\Wish\WishVoter;
 use App\Service\Wish\Subscription\WishSubscriptionService;
+use App\Validator\Wish\Subscription\WishSubscriptionValidator;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class WishSubscriptionController extends Controller
 {
     public function __construct(
+        private WishSubscriptionValidator $wishSubscriptionValidator,
         private WishSubscriptionService $wishSubscriptionService
     )
     {   
     }
 
-    #[Route('/list/{wishlist.id}/wish/{wish.id}/subscription/create', name: 'wish_subscription_create')]
-    public function create(Wishlist $wishlist, Wish $wish): Response
+    #[Route('/wish/{wish}/subscription/create', name: 'wish_subscription_create')]
+    public function create(Wish $wish, Request $request): Response
     {
-        // @todo check access to wishlist and wish
+        $this->denyAccessUnlessGranted(WishVoter::ACTION_CREATE_SUBSCRIPTION, $wish);
 
+        $user = $this->getUser();
+        $this->wishSubscriptionValidator->validateCreate($user, $wish);
+        
         $wish = $this->wishSubscriptionService->create(
             $wish,
-            $this->getUser()
+            $user
         );
 
-        return $this->render('wish/wish/index.html.twig', [ // @todo create template
-            'controller_name' => 'WishController',
-            'wish' => $wish
-        ]);
+        return $this->redirect($request->headers->get('referer'));
     }
 
-    #[Route('/list/{wishlist.id}/wish/{wish.id}/subscription/{wishSubscription.id}/remove', name: 'wish_subscription_remove')]
-    public function remove(Wishlist $wishlist, Wish $wish, WishSubscription $wishSubscription): Response
+    #[Route('/wish/{wish}/subscription/{wishSubscription}/remove', name: 'wish_subscription_remove')]
+    public function remove(Wish $wish, WishSubscription $wishSubscription, Request $request): Response
     {
-        // @todo check access to wishlist and wish
+        $this->denyAccessUnlessGranted(WishSubscriptionVoter::ACTION_REMOVE, $wishSubscription);
+        
         $this->wishSubscriptionService->remove(
             $wishSubscription
         );
 
-        return $this->render('wish/wish/index.html.twig', [ // @todo create template
-            'controller_name' => 'WishController',
-            'wish' => $wish
-        ]);
+        return $this->redirect($request->headers->get('referer'));
     }
 }
